Move feed page inside client component and add auth check
